@@ -30,64 +30,65 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.1.0",
-    "@angular/cdk": "13.1.0",
-    "@angular/common": "13.1.0",
-    "@angular/compiler": "13.1.0",
-    "@angular/core": "13.1.0",
-    "@angular/forms": "13.1.0",
-    "@angular/material": "13.1.0",
-    "@angular/platform-browser": "13.1.0",
-    "@angular/platform-browser-dynamic": "13.1.0",
-    "@angular/router": "13.1.0",
-    "@ngrx/component-store": "13.0.2",
-    "@ngrx/effects": "13.0.2",
-    "@ngrx/entity": "13.0.2",
-    "@ngrx/store": "13.0.2",
-    "@ngrx/store-devtools": "13.0.2",
-    "@nrwl/angular": "13.3.1",
+    "@angular/animations": "15.0.4",
+    "@angular/cdk": "15.0.4",
+    "@angular/common": "15.0.4",
+    "@angular/compiler": "15.0.4",
+    "@angular/core": "15.0.4",
+    "@angular/forms": "15.0.4",
+    "@angular/material": "15.0.4",
+    "@angular/platform-browser": "15.0.4",
+    "@angular/platform-browser-dynamic": "15.0.4",
+    "@angular/router": "15.0.4",
+    "@ngrx/component-store": "15.0.0",
+    "@ngrx/effects": "15.0.0",
+    "@ngrx/entity": "15.0.0",
+    "@ngrx/store": "15.0.0",
+    "@ngrx/store-devtools": "15.0.0",
+    "@nrwl/angular": "15.4.4",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
     "uuid": "^8.3.2",
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.0.4",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "13.0.4",
-    "@angular/compiler-cli": "13.1.0",
-    "@angular/language-service": "13.1.0",
+    "@angular-devkit/build-angular": "15.0.4",
+    "@angular-eslint/eslint-plugin": "15.0.0",
+    "@angular-eslint/eslint-plugin-template": "15.0.0",
+    "@angular-eslint/template-parser": "15.0.0",
+    "@angular/cli": "~15.0.0",
+    "@angular/compiler-cli": "15.0.4",
+    "@angular/language-service": "15.0.4",
     "@hirez_io/observer-spy": "^2.1.0",
-    "@nrwl/cli": "13.3.1",
-    "@nrwl/cypress": "13.3.1",
-    "@nrwl/eslint-plugin-nx": "13.3.1",
-    "@nrwl/jest": "13.3.1",
-    "@nrwl/linter": "13.3.1",
-    "@nrwl/tao": "13.3.1",
-    "@nrwl/workspace": "13.3.1",
-    "@types/jest": "27.0.2",
+    "@nrwl/cli": "15.4.4",
+    "@nrwl/cypress": "15.4.4",
+    "@nrwl/eslint-plugin-nx": "15.4.4",
+    "@nrwl/jest": "15.4.4",
+    "@nrwl/linter": "15.4.4",
+    "@nrwl/workspace": "15.4.4",
+    "@types/jest": "28.1.8",
     "@types/node": "14.14.33",
     "@types/uuid": "^8.3.3",
-    "@typescript-eslint/eslint-plugin": "5.3.1",
-    "@typescript-eslint/parser": "5.3.1",
-    "cypress": "^7.3.0",
+    "@typescript-eslint/eslint-plugin": "5.48.0",
+    "@typescript-eslint/parser": "5.48.0",
+    "cypress": "^11.0.0",
     "dotenv": "^10.0.0",
-    "eslint": "8.2.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.0.0",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "12.2.3",
     "json-server": "^0.16.3",
-    "ng-packagr": "^13.0.0",
+    "ng-packagr": "15.0.3",
+    "nx": "15.4.4",
     "postcss": "^8.3.9",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
-    "prettier": "^2.3.1",
-    "ts-jest": "27.0.5",
-    "ts-node": "~9.1.1",
-    "typescript": "4.4.4"
+    "postcss-import": "14.1.0",
+    "postcss-preset-env": "7.5.0",
+    "postcss-url": "10.1.3",
+    "prettier": "2.8.1",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4"
   }
 }
