@@ -2,4 +2,4 @@ import * as dotenv from 'dotenv';
 
 const plugin: Cypress.PluginConfig = (on, config) => {};
 
-module.exports = plugin;
+export default plugin;
